backend/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
reportlab
fpdf2
img2pdf
openpyxl
pillow
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
email-validator
python-pptx
pdfplumber
pdf2image
pypdf2
PyMuPDF
